- Moved the webapp files to cmd/app  - cmd/ contains the app and api folders.  - api will hold phone auth, and oauth api
