+++ Makefile
@@ -1,8 +1,15 @@
 tidy:
 	go mod tidy
 
-codegen:
+codegen: | bin
 	make -C openapi $@
 
+webapp: | bin
+	mkdir -p bin; go build -o bin/taxifyapp ./cmd/app
+
+bin:
+	mkdir -p bin
+
 clean:
+	rm -rf bin/*
 	rm -rf openapi/gen.go
