+++ cmd/app/route_auth.go
@@ -1,0 +1,69 @@
+package main
+
+import (
+	"net/http"
+
+	"kuberack.com/taxify/data"
+)
+
+// GET /login
+// Show the login page
+func login(writer http.ResponseWriter, request *http.Request) {
+	t := parseTemplateFiles("login.layout", "public.navbar", "login")
+	t.Execute(writer, nil)
+}
+
+// GET /signup
+// Show the signup page
+func signup(writer http.ResponseWriter, request *http.Request) {
+	generateHTML(writer, nil, "login.layout", "public.navbar", "signup")
+}
+
+// POST /signup
+// Create the user account
+func signupAccount(writer http.ResponseWriter, request *http.Request) {
+	err := request.ParseForm()
+	if err != nil {
+		danger(err, "Cannot parse form")
+	}
+	user := data.User{
+		Name:     request.PostFormValue("name"),
+		Email:    request.PostFormValue("email"),
+		Password: request.PostFormValue("password"),
+	}
+	if err := user.Create(); err != nil {
+		danger(err, "Cannot create user")
+	}
+	http.Redirect(writer, request, "/login", http.StatusFound)
+}
+
+func authenticate(w http.ResponseWriter, r *http.Request) {
+	r.ParseForm()
+	user, _ := data.UserByEmail(r.PostFormValue("email"))
+	if user.Password == data.Encrypt(r.PostFormValue("password")) {
+		session, err := user.CreateSession()
+		if err == nil {
+			cookie := http.Cookie{
+				Name:     "_cookie",
+				Value:    session.Uuid,
+				HttpOnly: true,
+			}
+			http.SetCookie(w, &cookie)
+			http.Redirect(w, r, "/", http.StatusFound)
+		}
+	} else {
+		http.Redirect(w, r, "/login", http.StatusFound)
+	}
+}
+
+// GET /logout
+// Logs the user out
+func logout(writer http.ResponseWriter, request *http.Request) {
+	cookie, err := request.Cookie("_cookie")
+	if err != http.ErrNoCookie {
+		warning(err, "Failed to get cookie")
+		session := data.Session{Uuid: cookie.Value}
+		session.DeleteByUUID()
+	}
+	http.Redirect(writer, request, "/", 302)
+}
