--- route_main.go
@@ -1,31 +1,0 @@
-package main
-
-import (
-	"net/http"
-
-	"kuberack.com/taxify/data"
-)
-
-// GET /err?msg=
-// shows the error message page
-func err(writer http.ResponseWriter, request *http.Request) {
-	vals := request.URL.Query()
-	_, err := session(writer, request)
-	if err != nil {
-		generateHTML(writer, vals.Get("msg"), "layout", "public.navbar", "error")
-	} else {
-		generateHTML(writer, vals.Get("msg"), "layout", "private.navbar", "error")
-	}
-}
-
-func index(w http.ResponseWriter, r *http.Request) {
-	threads, err := data.Threads()
-	if err == nil {
-		_, err := session(w, r)
-		if err != nil {
-			generateHTML(w, threads, "layout", "public.navbar", "index")
-		} else {
-			generateHTML(w, threads, "layout", "private.navbar", "index")
-		}
-	}
-}
